NeuroX Chat App/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
transformers
torch
scikit-learn
PyJWT
passlib[bcrypt]